medical-chatbot/medical-chatbot/requirements.txt
sentence-transformers
langchain-community
langchain
langchain-pinecone
pypdf
flask
flask-cors
streamlit
python-dotenv
together
requests
numpy
tqdm
pinecone
langchain-huggingface

